requirements.txt
pyrogram
tgcrypto
fastapi
uvicorn
motor
dnspython
python-dotenv